mdsim: fix FTBFS with gcc (GCC) 4.1.2 20080704 (Red Hat 4.1.2-48).
/theory2/u4/pcolberg/research/projects/halmd/src/halmd/mdsim/gpu/position/lattice.cpp: In member function ‘void halmd::mdsim::gpu::position::lattice<dimension, float_type>::set()’:
/theory2/u4/pcolberg/research/projects/halmd/src/halmd/mdsim/gpu/position/lattice.cpp:115: error: call of overloaded ‘pow(unsigned int&, int&)’ is ambiguous
/usr/include/bits/mathcalls.h:154: note: candidates are: double pow(double, double)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:361: note:                 long double std::pow(long double, int)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:357: note:                 float std::pow(float, int)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:353: note:                 double std::pow(double, int)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:349: note:                 long double std::pow(long double, long double)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:345: note:                 float std::pow(float, float)
/theory2/u4/pcolberg/research/projects/halmd/src/halmd/mdsim/gpu/position/lattice.cpp:119: error: call of overloaded ‘pow(unsigned int&, int&)’ is ambiguous
/usr/include/bits/mathcalls.h:154: note: candidates are: double pow(double, double)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:361: note:                 long double std::pow(long double, int)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:357: note:                 float std::pow(float, int)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:353: note:                 double std::pow(double, int)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:349: note:                 long double std::pow(long double, long double)
/usr/lib/gcc/x86_64-redhat-linux/4.1.2/../../../../include/c++/4.1.2/cmath:345: note:                 float std::pow(float, float)

+++ src/halmd/mdsim/gpu/position/lattice.cpp
@@ -112,11 +112,11 @@ void lattice<dimension, float_type>::set()
     // lower boundary for number of particles per lattice dimension
     unsigned int n = static_cast<unsigned int>(pow(particle->nbox / m, 1. / dimension));
     // lower boundary for total number of lattice sites
-    unsigned int N = m * static_cast<unsigned int>(pow(n, dimension));
+    unsigned int N = m * static_cast<unsigned int>(pow(static_cast<double>(n), dimension));
 
     if (N < particle->nbox) {
         n += 1;
-        N = m * static_cast<unsigned int>(pow(n, dimension));
+        N = m * static_cast<unsigned int>(pow(static_cast<double>(n), dimension));
     }
     if (N > particle->nbox) {
         LOG_WARNING("lattice not fully occupied (" << N << " sites)");
